requirements.txt
streamlit
streamlit-folium
folium
pandas
mysql-connector-python
psycopg2-binary
plotly